Rename lib/ into src/
