+++ src/__fixtures__/.editorconfig
