+++ src/__fixtures__/endofline.cr.fixture
