+++ src/__fixtures__/endofline.mixed.fixture
