+++ src/__fixtures__/newlines.endoffile.invalid.less.fixture
