+++ src/__fixtures__/newlines.endoffile.singleline.fixture
