+++ src/__fixtures__/newlines.endoffile.valid.fixture
