+++ src/__fixtures__/trailingspaces.ignores.fixture
