+++ src/__fixtures__/trailingspaces.invalid.fixture
