+++ src/__fixtures__/trailingspaces.valid.fixture
