+++ src/constants/editorconfig-mappings.js
